feat(ssh): configure SK backed ssh keys
_sk ssh keys are stub files which can be used with the corresponding
security key to derive the actual private key. This allows us to
securely version control them, enabling automatic distribution and
configuration on all machines.

Since we are using multiple security keys for redundancy and security
purposes this means that we need to deal with alternative identities for
most ssh targets and cannot simply rely on the "IdentityFile" directive
in .ssh/config match blocks. This is accomplished by loading the private
keys into ssh-agent which can select keys based on the presence of
physical security keys. Unfortunately neither nix nor home-manager offer
an option to declaratively add specific keys to the ssh-agent, since
this is an interactive operation for non-sk ssh keys. To work around
this issue, this commit adds code to .zshrc which will automatically add
all keys ending in "_sk" at shell startup.