fix(zsh): update zsh-autocomplete config for new version
In commit 10494fccd12989a72fbd027fe7a96878f6262def zsh-autocomplete
changed the way in which certain completions, especially
insert-unambigous and tab behaviour are configured. Since plugin version
was not pinned in zplug, at some point we automatically updated past
this commit, breaking existing behaviour.

This commit changes the relevant zstyle and bindkey commands to the new
configuration options and also pins the version of zsh-autocomplete for
zplug to load. There remains to be found a good way of automatically,
but auditably updating the pinned version of such plugins to avoid
plugins going stale. To this end a TODO file has been added to track the
necessity