refactor(ausweisapp): disable autostart
Currently there is no way of launching just the Ausweisapp background
service without popping up the UI window. This window then steals screen
real estate from the apps we actually want to launch.

In a future, with an autostart system that is integrated with the window
manager, it might be possible to lauch the Ausweisapp window on an out of the
way workspace or immediately close it after opening. At this point the
Ausweisapp should be re-added to the autostart since launching the
service at startup does increase useability. Until then, autostarting
the app is more detrimental than helpful