feat(nvidia): add gwe to autostart for fancontrol
Nvidia GPUs don't have persistent fan profiles beside their VBIOS
controlled ones. It is generally desirable to alter the default fan
curve to achieve quiet operation within the cooling capabilities of
one's hardware.

Green with Envy is a software that can be used to control the fan curve
(among other things) of Nvidia GPUs. However it needs to be running in
order to continuously adapt the fan speed based on temperature
information.

This commit adds the gwe package to the user packages and adds an entry
to the autostart script which spawns the process, but does not pop up
the app window to not clutter the screen on start. The gwe ui can be
accessed from the bar widget when needed.