fix(xmonad): systemd user services have no access to PATH
Many graphical services are managed by systemd and therefore inherit the
systemd user session. An important set of these services are the
xdg-portal* services which are responsible for dispatching default
applications for opening URIs or files. Without knowledge of the system
the portal can determine but not reach the default applications, leading
xdg-open to fail silently and gio mime to display "No Apps available"
(e.g. https://github.com/flatpak/xdg-desktop-portal-gtk/issues/440)

Propagating the values to the systemd user session is individual to the
WM (and modules or packaged used to set it up, with some implementing
partial fixes). By default, home-manager propagates all necessary XDG_
env vars in its xsession, but not PATH. This is not fixed inside the
xmonad module either (as of writing).

This commit adds "PATH" to the propagated env vars, when enabling
xmonad. Other WM module might not need this correction