feat(user): Add nixGL wrapper to fix openGL issues on non-NixOS
While OpenGL presents a seemingly unified API for applications, the way
the calls to the GPU are dispatched includes a non-deterministic
selection of vendor driver modules at runtime which makes it very hard
to properly encapsulate in nix. While NixOS includes a solution which
moves the variability into the system config and away from the
individual GL application packages, this is not possible on NixOS
systems, leading to broken GL apps when using nix/home-manager on other
distros.
For more details, refer to https://alternativebit.fr/posts/nixos/nix-opengl-and-ubuntu-integration-nightmare/

There is a en effort called nixGL which wraps pkgs at build time to
include references to a specific openGL driver. This is a suboptimal
solution, since for any non-free (nvidia) drivers it needs to query the
installed driver version to avoid ABI-incompatibility. For now, we only
include the configuration linking against the mesa opengl library, which
should also be available on nvidia installations and can be included in
a pure fashion.

Wrapping for propietary nvidia in standalone or bumblebee/optimus setups
needs to be investigated further