refactor(cuda): replace global cudaSupport with opt-in
Currently we are enabling cudaSupport in the global config, which causes
all transitively included packages with cudaSupport to be rebuilt
(because cuda is unfree and therefore complicated to redistribute via
cache servers). Often this rebuilts a lot of dependencies which are only
used by non-cuda top-level packages and therefore wastes a lot of time
and cpu-cycles on every flake update.

This commit instead defines a new global option "_cudaSupportAvailable"
in the nvidia system module and enables cudaSupport directly on the
relevant packages (currently blender). The new option allows modules to
configure themselves based on availability of cuda if they always want
to include cudaSupport wherever possible