feat(work/git): add `rip` (rebase in place) alias
When working to prepare with complex pull requests that require fixups
to multiple commits, it can become very tedious to fix the merge
conflicts that occur when unthreading and re-orderig even fixup commits.
In such situations it would be desirable to locally rebase often to keep
the interweaving of fixups to an acceptable level. When doing this, it
is often good to split the tracking of the target branch and cleaning up
the fixups into two phases by first rebasing the fixups "in-place" (i.e.
without changing the merge-base of the current branch with the target
branch) and only then rebasing the cleaned changeset on top of the
current tip of the target branch.

This commit adds an alias to speed up this first operation. The alias is
called "rip" for rebase in place. It takes a single positional
argument naming the target branch (default: "origin/HEAD" for trunk
based workflows), which it used to calculate the merge base of the
current head and the target branch and then launching an interactive,
auto-stashing, autosquashing rebase on this merge base