feat(zsh): add BronzeDeer/zsh-completion-sync
I wrote this plugin to get around the problem of direnv shells not
reloading the autocompletion for added fpaths (as well as nix-shells
"only" putting includes on XDG_DATA_DIRS and not on the fpath. This
plugin fixes both and dynamically reloads the zsh autocomplete system.
Further, it also includes some basic compatibility to reload
zsh-autocomplete which disable the normal compinit function