feat(workstation): make display config independent of xrandr names
autorandr configures monitors via their edid, however by default it
still matches on xrandr names (e.g. HDMI-1, DP-4) which are not
necessarily stable or easy to predict. This means that after replugging
monitors or a kernel or driver update the same configuration might not
work anymore.

Autorandr can be put into a "matchEdid" mode which will dynamically
rename the dispaly identifiers in the config to match current xrandr
names, based on the edid values. For many people this works flawlessly,
however for the current workstation config it failed, since the outer
monitors shared timing information and the serial matching mode was
broken.

This commit overrides the autorandr pkg with a fixed version from an
open PR, activates the matchEdid mode on the autorandr service, and
renames the monitor names in the configuration to something more
semantically meaningful (left, middle, right)