feat: use home-manager config via hm module on nixos hosts
In some cases it is desirable to have the home-configuration of a user
to react to certain facts of the host on which they will be deployed.
One such example is deactivating nixGL wrapping when running on NixOS.
However detecting facts about the host at build time via execs etc would
break purity and a lot of the benefits of the nix-way of managing
configurations. An alternative is to define additional module inputs to
communicate all relevant facts, however this can quickly become messy
and hard to maintain, bloating flake.nix with lots of plumbing code.

The better alternative to this is using the home-manager nixos module,
which exposes the osConfig as an additional standardized argument. This
allows all home-manager modules to track against the well established
nixos config structure as a unifying api for passing additional
information. However additional care has to be taken to keep the
home-manager modules useable as standalone flake outputs via
homeConfigurations.<user>.

This commit creates a new directory "users" which will store the entry
point for each user module. For the existing joel user this also means
moving the module imports into the entrypoint instead of the inputs of
homeManagerConfiguration. The user module can then be imported by both
homeConfiguration and home-manager.users.<user>. Note that
home-manager's "extraSpecialArgs" need to be defined in both places via
slightly different keys.

With this refactoring done, future commits can add special case handling
on nixOS by checking for the presence of non-null osConfig inside the
module inputs. it also paves the way for reintegrating the separate work
config branch whithout maintaining so much duplicate code