feat(pay-respects): add nix-index for command-not-found functionality
nix-index is a package that builds/uses and index of all files in all of
nixpkgs. This allows it to easily provide command-not-found
functionality, which among other things, pay-respects relies on.

The pay-respects module will now enable nix-index via home-manager and
by default integrate it with all enabled, home-manager manged shells.

It also checks the modules for the presence of an input called
`nix-index-database` (https://github.com/nix-community/nix-index-database)
and uses this precompiled nix-index database if present in the inputs.

Furthermore this commit also adds the nix-index-database as flake input
and makes it available to the two current usages of home-manager
profiles