feat(workstation): clean /tmp on boot
Note that on nixos by default /tmp is not on a tmpfs to avoid large nix
builds crashing from too small tmpfs. Disk-churn is not a significant
concern, but clutter in /tmp is, so it is enough to empty it on next
boot. (And cold boot attacks can recover tmpfs anyway)