fix(personal/git): stop "reuse" alias clobbering last commit
The reuse alias erroneously included --amend, which would integrate the
new changes for which a commit message should be reused into the last
commit. The effect of this can be seen in commit e0742e5 which absorbed
the "remove zplug" commit. Sadly this commit is already on master.

This change should stop this problem happening again in the future