feat(kitty): improve symbol support with explicit mapping/fallback
kitty has a explicit font-fallback and symbol mapping features, which
allows it to no be reliant on patched fonts (which can sometimes lead to
weird issues).

This commit configures kitty to use the un-patched
version of the currently configured font (fira-code) and install the
NerdFonts symbols only font, explicitly mapping it's symbols in kitty's
settings. Other applications implementing font-fallback should now also
have an easier time accessing nerd-font symbols even if the configured
font is not a patched font