fix(hyprlock): correctly declare animation time in config
Currently, hypridle bugs out, always triggering its auto lock after 30
minutes without registering non-idle states. This seems to be partially
connected to the fade in and out animations for hyprlock being declared
based on an example with a deprecated config format.

This commit correctly declares fade in and out as a 5 second transition
to the lock screen and a half second transition back to the desktop
after unlocking