fix(laptop/sudo): yubikey auth doesn't work with different hostname
They keyid used by the pam_u2f module is dependent on the "origin"
value, which by default is the dependent on the hostname. The existing
u2f-keys were generated with the origin value defaulting to
"pam://nixos-workstation".

This commit hardcodes this origin to ensure that the u2f_keys file works
across all machines